Small changes on the form that creates markers and moved mapview to features
