+++ webapp/src/app/store.ts
@@ -1,6 +1,6 @@
 import { configureStore, ThunkAction, Action } from '@reduxjs/toolkit';
 import counterReducer from '../features/counter/counterSlice';
-import mapReducer from '../mapView/MapViewSlice'
+import mapReducer from '../features/mapView/MapViewSlice'
 
 export const store = configureStore({
   reducer: {
